优化宠物喂食机和加热器

+++ app/src/main/java/com/github/miwu/miot/widget/AirPurifierBar.kt
@@ -67,7 +67,7 @@ class AirPurifierBar(context: Context) : MiotBaseWidget<Binding>(context) {
 
             "mode" -> {
                 modeValue = 0
-                refreshMode(value as Int)
+                refreshMode((value as Double).toInt())
             }
         }
     }
