+++ app/src/main/java/com/github/miwu/miot/widget/HeaterBar.kt
@@ -67,7 +67,7 @@ class HeaterBar(context: Context) : MiotBaseWidget<Binding>(context) {
 
             "mode" -> {
                 modeValue = 0
-                refreshMode(value as Int)
+                refreshMode((value as Double).toInt())
             }
         }
     }
